implement lists in the same files as singular data
