+++ lib/src/lib.rs
@@ -11,21 +11,15 @@ pub mod replica;
 pub mod result;
 pub mod status;
 pub mod string;
-pub mod stringlist;
 pub mod task;
-pub mod tasklist;
 pub mod uuid;
-pub mod uuidlist;
 
 pub(crate) mod types {
     pub(crate) use crate::annotation::{TCAnnotation, TCAnnotationList};
     pub(crate) use crate::replica::TCReplica;
     pub(crate) use crate::result::TCResult;
     pub(crate) use crate::status::TCStatus;
-    pub(crate) use crate::string::TCString;
-    pub(crate) use crate::stringlist::TCStringList;
-    pub(crate) use crate::task::TCTask;
-    pub(crate) use crate::tasklist::TCTaskList;
-    pub(crate) use crate::uuid::TCUuid;
-    pub(crate) use crate::uuidlist::TCUuidList;
+    pub(crate) use crate::string::{TCString, TCStringList};
+    pub(crate) use crate::task::{TCTask, TCTaskList};
+    pub(crate) use crate::uuid::{TCUuid, TCUuidList};
 }
